Refactored tests into two files
