--- promise-cacheable-tests.ts
@@ -1,4 +1,0 @@
-// export const promiseCacheableTests = (Service, Service2) => {
-
-  
-// }
